+++ CHANGELOG.md
@@ -26,11 +26,13 @@
     - Sky - Beedle's Crystals
     - Sealed Grounds - Gorko's Goddess Wall Reward
   - The Sea Chart can now be hinted when Sandship is a required dungeon
+  - Improved how the randomizer categorizes regions as barren
 ### Bugfixes
 - Added some tricks that were in the logic files but not in options (by NULL)
 - Fix bug with arc-replacements causing an error if the arc doesn't exist in ObjectPack, it's ignored now
 - Fix Sandship Timeshift stone not being unset when leaving the dungeon in entrance rando
 - Fix a bug in trials where you could be unable to get the item with glitches
+- Fix a bug which prevented hints from being placed correctly 
 
 ## 1.1.0
 ### Options